# .well-known/openapi.yaml
openapi: 3.0.1
info:
    title: 单词本
    description: 个性化的英文单词本，可以增加、删除和浏览单词本中的单词，背单词时从已有单词本中随机抽取单词生成句子或者段落。
    version: "v1"
servers:
    - url: http://127.0.0.1:8081
paths:
    /read_board:
        post:
            operationId: readBoard
            summary: 从远程文件中读取棋盘
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/addBoard"
            responses:
                "200":
                    description: 远程文件中的棋盘读取成功
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/messageResponse"

    /get_wordbook:
        get:
            operationId: getWordbook
            summary: 展示单词列表
            responses:
                "200":
                    description: 列表展示完成
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/getWordbook"
    /generate_sentences:
        post:        
            operationId: generateSentences
            summary: 背单词，生成句子
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/generateSentences"
            responses:
                "200":
                    description: 句子生成成功
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/responseSentences"
    /add_word:
        post:
            operationId: addWord
            summary: 在单词本中添加一个单词
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/addWord"
            responses:
                "200":
                    description: 单词添加成功
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/messageResponse"
    /delete_word:
        delete:
            operationId: deleteWord
            summary: 从单词本中删除一个单词
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/deleteWord"
            responses:
                "200":
                    description: 单词删除成功
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/messageResponse"

components:
    schemas:
        addBoard:
            type: object
            required: [ url ]
            properties:
                url:
                    type: string
                    description: 远程文件url地址，用<url><\url>包含
        getWordbook:
            type: object
            required: [wordbook]
            properties:
                wordbook:
                    type: array
                    items:
                        type: string
                    description: 单词本单词列表
        addWord:
            type: object
            required: [word]
            properties:
                word:
                    type: string
                    description: 需要添加到单词本中的一个单词
        deleteWord:
            type: object
            required: [word]
            properties:
                word:
                    type: string
                    description: 需要删除的单词
        generateSentences:
            type: object
            required: [word_number]
            properties:
                word_number:
                    type: integer
                    description: 几个单词
        responseSentences:
            type: object
            required: [words]
            properties:
                words:
                    type: array
                    items:
                        type: string
                        description: 随机抽取的英文单词
        messageResponse:
            type: object
            required: [message]
            properties:
                message:
                    type: string
                    description: 回复信息